feat: Add compiler, editor, and workspace commands
This commit adds the following commands to the application:

- `compile`: Compiles Typst source code.
- `get_cursor_position`: Gets the cursor position in the document.
- `render`: Renders the current document.
- `render_page`: Renders a single page of the document.
- `compile_file`: Compiles a file and emits results.
- `page_click`: Handles a click on a page.
- `open_workspace`: Opens a workspace.
- `open_file`: Opens a file in the workspace.
- `export_to`: Exports the compiled document.
- `autocomplete`: Provides autocomplete suggestions.
- `tooltip`: Provides tooltip information.
