+++ src-tauri/src/commands/mod.rs
@@ -1,0 +1,3 @@
+pub mod compiler;
+pub mod editor;
+pub mod workspace;
